services/rider-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
requests==2.31.0
APScheduler==3.10.4
